dSEIR: R package for dynamic SEIR model
For the prediction of the epidemic trend of COVID-19 pneumonia, a dynamic SEIR model was proposed by Kuangnan Fang , Rui Ten et al., which is based on the Susceptible-Exposed-Infected-Removed model of transmission dynamics. The advantages of the model are as follows :(1) the model parameters can be dynamically estimated based on the real-time updated data; (2) compared with the traditional SEIR model, which assumes a closed environment, the proposed model can take into account the impact of population flow on disease transmission and has a better prediction effect.